+++ mantrap/agents/integrator.py
@@ -18,10 +18,10 @@ class IntegratorDTAgent(LinearAgent):
                           [0, 1, 0, 0, 0],
                           [0, 0, 0, 0, 0],
                           [0, 0, 0, 0, 0],
-                          [0, 0, 0, 0, 1]]).float()
+                          [0, 0, 0, 0, 1]])
         B = torch.tensor([[dt, 0, 1, 0, 0],
-                          [0, dt, 0, 1, 0]]).t().float()
-        T = torch.tensor([0, 0, 0, 0, dt]).t().float()
+                          [0, dt, 0, 1, 0]]).t()
+        T = torch.tensor([0, 0, 0, 0, dt]).t()
         return A, B, T
 
     @staticmethod
